Postorder traverse tree when renaming pages
Given an old_tree: A <- B <- C
And   a  new_tree: A <- C <- B

re_move_pages() created a cycle if it tried to move B to a subpage of C
first. The new implementation does a postorder traversal to make sure
this does not happen. For this the parent list old_tree needs to be
converted to a child list reverse_tree.